add util function to populate html email template with data
