+++ src/lib/utils.ts
@@ -53,3 +53,20 @@ export function trapFocus(node: HTMLElement) {
     },
   }
 }
+        
+// replace html template with data
+export function addDataToHtmlTemplate(html, template) {
+  const htmlBody = html.replace(/{{(.*?)}}/g, (_, key) => {
+    const keys = key.trim().split('.');
+    let value = template.data;
+    for (const k of keys) {
+      value = value[k];
+      if (value === undefined) {
+        return '';
+      }
+    }
+    console.log(value);
+    return value;
+  });
+  return htmlBody;
+}
