+++ src/routes/api/action/+server.ts
@@ -6,6 +6,7 @@ import postmark from 'postmark'
 import {
   POSTMARK_API_TOKEN,
 } from '$env/static/private'
+import { addDataToHtmlTemplate } from '$lib/utils'
 
 
 export const POST: RequestHandler = async ({ request, locals }) => {
@@ -95,23 +96,8 @@ export const POST: RequestHandler = async ({ request, locals }) => {
       // get html template from firebase
       const document = await adminDb.collection('templates').doc(template.name).get()
 
-      const html = document.data()?.html
-
-      // replace html template with data
-      const htmlBody = html.replace(/{{(.*?)}}/g, (_, key) => {
-        const keys = key.trim().split('.');
-        let value = template.data;
-        for (const k of keys) {
-          value = value[k];
-          if (value === undefined) {
-            return '';
-          }
-        }
-        return value;
-      });
-
-
-
+      const htmlBody = addDataToHtmlTemplate(document.data()?.html, template);
+      
       const emailData: Data.EmailData = {
         From: 'donotreply@gbstem.org',
         To: to,
