+++ src/routes/api/interview/+server.ts
@@ -6,6 +6,7 @@ import type { FirebaseError } from 'firebase-admin'
 import {
   POSTMARK_API_TOKEN,
 } from '$env/static/private'
+import { addDataToHtmlTemplate } from '$lib/utils'
 
 export const POST: RequestHandler = async ({ request, locals }) => {
   let topError
@@ -36,21 +37,7 @@ export const POST: RequestHandler = async ({ request, locals }) => {
         }
         const document = await adminDb.collection('templates').doc(template.name).get()
     
-        const html = document.data()?.html
-    
-        // replace html template with data
-        const htmlBody = html.replace(/{{(.*?)}}/g, (_, key) => {
-          const keys = key.trim().split('.');
-          let value = template.data;
-          for (const k of keys) {
-            value = value[k];
-            if (value === undefined) {
-              return '';
-            }
-          }
-          console.log(value);
-          return value;
-        });
+        const htmlBody = addDataToHtmlTemplate(document.data()?.html, template);
     
         const emailData: Data.EmailData = {
           From: 'donotreply@gbstem.org',
